requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cloudinary==1.28.0
dj-database-url==0.5.0
Django==4.2.1
django-admin-interface==0.26.0
django-ckeditor==6.5.1
django-cleanup==7.0.0
django-colorfield==0.5.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-js-asset==1.2.2
gunicorn==20.1.0
Pillow==9.5.0
psycopg2==2.9.2
python-decouple==3.8
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
whitenoise==5.3.0
